# projects/sbos-minimal/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Brick Server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/brickapi/v1/entities/upload": {
      "post": {
        "tags": [
          "Entities",
          "Entities"
        ],
        "summary": "Uplaod a Turtle file",
        "description": "Upload a Turtle file. An example file: https://gitlab.com/jbkoh/sbos-dev/blob/dev/examples/data/bldg.ttl",
        "operationId": "upload_brickapi_v1_entities_upload_post",
        "parameters": [
          {
            "description": "The name of the graph. This is similar to a database name in relational databases.",
            "required": false,
            "schema": {
              "title": "Named Graph",
              "type": "string",
              "description": "The name of the graph. This is similar to a database name in relational databases."
            },
            "name": "named_graph",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_brickapi_v1_entities_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsSuccess"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/brickapi/v1/entities/": {
      "get": {
        "tags": [
          "Entities",
          "Entities"
        ],
        "summary": "Get Entity By Id",
        "description": "Get information about an entity including type and its relationships with others. The definition of entity: An entity can be defined in two ways. It's an instance of a (Brick) Class. More specifically, it is either a physical or a virtual thing whose properties are well-maintained to be a thing. Top three Brick Classes are Point (e.g., sensors, setpoints, etc.), Equipment (e.g., VAV, Luminaire, AHU, etc.), and Location (e.g., Room, Floor, etc.)",
        "operationId": "get_entity_by_id_brickapi_v1_entities__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            },
            "name": "entity_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/brickapi/v1/entities/list": {
      "post": {
        "tags": [
          "Entities",
          "Entities"
        ],
        "summary": "Post",
        "description": "List all entities with their types and relations.",
        "operationId": "post_brickapi_v1_entities_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Params",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ListEntityParams"
                  }
                ],
                "description": "A list of object URIs for the corresponding predicate. Brick Server will find entities having relations with all the objects with the predicate (i.e., AND operation.)",
                "default": {
                  "hasPoint": [],
                  "isPointOf": [],
                  "hasPart": [],
                  "isPartOf": [],
                  "hasLocation": [],
                  "isLocationOf": [],
                  "feeds": [],
                  "isFedBy": []
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityIds"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_brickapi_v1_entities_upload_post": {
        "title": "Body_upload_brickapi_v1_entities_upload_post",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "Entity": {
        "title": "Entity",
        "required": [
          "type",
          "entity_id"
        ],
        "type": "object",
        "properties": {
          "relationships": {
            "title": "Relationships",
            "type": "array",
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "The list of relationships for the target entity. Assuming the target entity is the subject, each relation consists of the subject's predicate and object.s",
            "default": []
          },
          "type": {
            "title": "Type",
            "type": "string",
            "description": "The entity's type, which is often a Brick Class."
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string",
            "description": "The identifier of an entity. Often a URI. This should be unique across the target systems (i.e., the graphs of the interest.)"
          },
          "name": {
            "title": "Name",
            "type": "string",
            "description": "An informative name for the entity. This does not have to be unique."
          }
        }
      },
      "EntityIds": {
        "title": "EntityIds",
        "required": [
          "entity_ids"
        ],
        "type": "object",
        "properties": {
          "entity_ids": {
            "title": "Entity Ids",
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of `entity_id`s"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "IsSuccess": {
        "title": "IsSuccess",
        "type": "object",
        "properties": {
          "is_success": {
            "title": "Is Success",
            "type": "boolean",
            "default": true
          },
          "reason": {
            "title": "Reason",
            "type": "string",
            "default": ""
          }
        }
      },
      "ListEntityParams": {
        "title": "ListEntityParams",
        "type": "object",
        "properties": {
          "hasPoint": {
            "title": "Haspoint",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "isPointOf": {
            "title": "Ispointof",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "hasPart": {
            "title": "Haspart",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "isPartOf": {
            "title": "Ispartof",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "hasLocation": {
            "title": "Haslocation",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "isLocationOf": {
            "title": "Islocationof",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "feeds": {
            "title": "Feeds",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "isFedBy": {
            "title": "Isfedby",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}